Divide `open-street-maps` into a library and binary
Going to work on more examples using the OSM `TileFetcher` and want to keep everything in the same examples directory. Changing this enables that choice.
