+++ examples/open-street-maps/Cargo.toml
@@ -13,3 +13,7 @@ geo = { workspace = true }
 image = { workspace = true }
 reqwest = { version = "0.12.7", features = ["blocking"] }
 snapper = { path = "../../snapper" }
+
+[[bin]]
+path = "src/point/main.rs"
+name = "point"
