+++ examples/open-street-maps/src/point/main.rs
@@ -1,7 +1,4 @@
-use std::io::Cursor;
-
-use image::{DynamicImage, ImageFormat, ImageReader};
-use reqwest::blocking::ClientBuilder;
+use open_street_maps::tile_fetcher;
 use snapper::SnapperBuilder;
 
 fn main() -> Result<(), anyhow::Error> {
@@ -34,28 +31,4 @@ fn main() -> Result<(), anyhow::Error> {
         .save("example.png")?;
 
     Ok(())
-}
-
-fn tile_fetcher(x: i32, y: i32, zoom: u8) -> Result<DynamicImage, snapper::Error> {
-    let address = format!("https://a.tile.osm.org/{zoom}/{x}/{y}.png");
-
-    let client = ClientBuilder::new()
-        .user_agent("snapper / 0.1.0")
-        .build()
-        .map_err(anyhow::Error::from)?;
-
-    let cursor = client
-        .get(&address)
-        .send()
-        .and_then(|response| response.error_for_status())
-        .and_then(|response| response.bytes())
-        .map(|bytes| Cursor::new(bytes))
-        .map_err(anyhow::Error::from)?;
-
-    let mut image_reader = ImageReader::new(cursor);
-    image_reader.set_format(ImageFormat::Png);
-
-    let image = image_reader.decode().map_err(anyhow::Error::from)?;
-
-    Ok(image)
 }
